requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
cryptography==35.0.0
cssselect==1.1.0
fake-useragent==0.1.11
idna==2.10
lxml==4.6.3
Naked==0.1.31
parse==1.19.0
pdfminer.six==20211012
pycparser==2.21
pycryptodome==3.10.1
pyee==8.1.0
PyPDF2==1.26.0
pyppeteer==0.2.5
pyquery==1.4.3
PyYAML==5.4.1
requests==2.25.1
requests-html==0.10.0
shellescape==3.8.1
six==1.15.0
soupsieve==2.2.1
tqdm==4.60.0
urllib3==1.26.4
w3lib==1.22.0
websockets==8.1
